+++ test/test_grayscale.rb
@@ -39,6 +39,16 @@ module TestColor
       assert_in_tolerance(0.363, @gs.lighten_by(10).g)
     end
 
+    def test_inspect
+      assert_equal("Grayscale [33.00%]", @gs.inspect)
+    end
+  end
+
+  class TestGrayscaleConversions < Minitest::Test
+    def setup
+      @gs = Color::Grayscale.from_percentage(33)
+    end
+
     def test_to_cmyk
       cmyk = @gs.to_cmyk
       assert_kind_of(Color::CMYK, cmyk)
@@ -46,11 +56,12 @@ module TestColor
       assert_in_tolerance(0.0, cmyk.m)
       assert_in_tolerance(0.0, cmyk.y)
       assert_in_tolerance(0.67, cmyk.k)
+      assert_equal(Color::CMYK.from_percentage(0, 0, 0, 100), Color::Grayscale.from_percentage(0).to_cmyk)
     end
 
     def test_to_grayscale
       assert_equal(@gs, @gs.to_grayscale)
-      assert_equal(@gs, @gs.to_grayscale)
+      assert_kind_of(Color::Grayscale, @gs.to_grayscale)
     end
 
     def test_to_hsl
@@ -59,6 +70,16 @@ module TestColor
       assert_in_tolerance(0.0, hsl.h)
       assert_in_tolerance(0.0, hsl.s)
       assert_in_tolerance(0.33, hsl.l)
+      assert_equal(Color::HSL.from_values(0, 0, 0), Color::Grayscale.from_percentage(0).to_hsl)
+    end
+
+    def test_to_lab
+      lab = @gs.to_lab
+      assert_kind_of(Color::CIELAB, lab)
+      assert_in_tolerance(35.78746, lab.l)
+      assert_in_tolerance(0.0031199, lab.a)
+      assert_in_tolerance(-0.000639, lab.b)
+      assert_equal(Color::CIELAB.from_values(0, 0, 0), Color::Grayscale.from_percentage(0).to_lab)
     end
 
     def test_to_rgb
@@ -67,6 +88,16 @@ module TestColor
       assert_in_tolerance(0.33, rgb.r)
       assert_in_tolerance(0.33, rgb.g)
       assert_in_tolerance(0.33, rgb.b)
+      assert_equal(Color::RGB.from_values(0, 0, 0), Color::Grayscale.from_percentage(0).to_rgb)
+    end
+
+    def test_to_xyz
+      xyz = @gs.to_xyz
+      assert_kind_of(Color::XYZ, xyz)
+      assert_in_tolerance(0.08457, xyz.x)
+      assert_in_tolerance(0.08898, xyz.y)
+      assert_in_tolerance(0.09688, xyz.z)
+      assert_equal(Color::XYZ.from_values(0, 0, 0), Color::Grayscale.from_percentage(0).to_xyz)
     end
 
     def test_to_yiq
@@ -75,10 +106,11 @@ module TestColor
       assert_in_tolerance(0.33, yiq.y)
       assert_in_tolerance(0.0, yiq.i)
       assert_in_tolerance(0.0, yiq.q)
+      assert_equal(Color::YIQ.from_values(0, 0, 0), Color::Grayscale.from_percentage(0).to_yiq)
     end
 
-    def test_inspect
-      assert_equal("Grayscale [33.00%]", @gs.inspect)
+    def test_to_internal
+      assert_equal([0.0], Color::Grayscale.from_percentage(0).to_internal)
     end
   end
 end